Backend/requirements.txt
pymongo==3.12.1
mongoengine==0.23.1
django==3.2.8
djangorestframework==3.12.4
drf-spectacular # for api documentation
django-cors-headers